Reorg helpers in acceptance package.
